+++ source/ProgArgs.cpp
@@ -551,6 +551,11 @@ void ProgArgs::defineAllowedArgs()
             "\"--" ARG_S3BUCKETTAG_LONG "\")")
 /*s3b*/	(ARG_S3BUCKETVER_LONG, bpo::bool_switch(&this->doS3BucketVersioning),
             "Activate bucket versioning operations.")
+		(ARG_S3LISTMPU_LONG, bpo::bool_switch(&this->doS3ListMPU),
+			"Enable list multipart uploads requests during multipart "
+			"uploads. (Off by default)")
+		(ARG_S3LISTPARTS_LONG, bpo::bool_switch(&this->doS3ListParts),
+			"Enable list parts requests during multipart uploads. (Off by default)")
 /*s3b*/	(ARG_S3BUCKETVERVERIFY_LONG, bpo::bool_switch(&this->doS3BucketVersioningVerify),
             "Verify the correctness of S3 bucket versioning settings. (Requires "
             "\"--" ARG_S3BUCKETVER_LONG "\")")
@@ -891,6 +896,8 @@ void ProgArgs::defineDefaults()
     this->doS3BucketVersioningVerify = false;
     this->doS3ObjectTag = false;
     this->doS3ObjectTagVerify = false;
+    this->doS3ListMPU = false;
+    this->doS3ListParts = false;
     this->doS3ObjectLockCfg = false;
     this->doS3ObjectLockCfgVerify = false;
 	this->useOpsLogLocking = false;
@@ -3329,6 +3336,8 @@ void ProgArgs::setFromPropertyTreeForService(bpt::ptree& tree)
     doS3BucketTag = tree.get<bool>(ARG_S3BUCKETTAG_LONG);
     doS3BucketTagVerify = tree.get<bool>(ARG_S3BUCKETTAGVERIFY_LONG);
     doS3BucketVersioning = tree.get<bool>(ARG_S3BUCKETVER_LONG);
+    doS3ListMPU = tree.get<bool>(ARG_S3LISTMPU_LONG);
+    doS3ListParts = tree.get<bool>(ARG_S3LISTPARTS_LONG);
     doS3BucketVersioningVerify = tree.get<bool>(ARG_S3BUCKETVERVERIFY_LONG);
     doS3ObjectTag = tree.get<bool>(ARG_S3OBJTAG_LONG);
     doS3ObjectTagVerify = tree.get<bool>(ARG_S3OBJTAGVERIFY_LONG);
@@ -3556,6 +3565,8 @@ void ProgArgs::getAsPropertyTreeForService(bpt::ptree& outTree, size_t serviceRa
     outTree.put(ARG_S3BUCKETTAG_LONG, doS3BucketTag);
     outTree.put(ARG_S3BUCKETTAGVERIFY_LONG, doS3BucketTagVerify);
     outTree.put(ARG_S3BUCKETVER_LONG, doS3BucketVersioning);
+    outTree.put(ARG_S3LISTMPU_LONG, doS3ListMPU);
+    outTree.put(ARG_S3LISTPARTS_LONG, doS3ListParts);
     outTree.put(ARG_S3BUCKETVERVERIFY_LONG, doS3BucketVersioningVerify);
     outTree.put(ARG_S3OBJTAG_LONG, doS3ObjectTag);
     outTree.put(ARG_S3OBJTAGVERIFY_LONG, doS3ObjectTagVerify);